+++ main/CMakeLists.txt
@@ -1,14 +1,12 @@
 idf_component_register(SRCS "main.c"
                             "bridge.c"
                             "ethernet.c"
-                            "auth_session.c"
-                            "auth_handlers.c"
+                            "auth.c"
                             "http_server.c"
                             "info_handlers.c"
                             "nv_storage.c"
                             "serial.c"
                             "setting_items.c"
-                            "ssdp.c"
                             "tcp_client.c"
                             "tcp_server.c"
                             "wifi_apsta.c"
