--- main/ssdp.c
@@ -1,1 +1,0 @@
-#include "ssdp.h"
