--- main/ssdp.h
@@ -1,3 +1,0 @@
-#pragma once
-
-typedef int* ssdp_config_t;
