+++ main/wifi_scan.c
@@ -1,6 +1,7 @@
 #include "wifi_scan.h"
 #include "json_utils.h"
-#include "auth_handlers.h"
+#include "auth.h"
+
 #include <esp_log.h>
 #include <esp_wifi.h>
 #include <freertos/FreeRTOS.h>
@@ -9,7 +10,9 @@
 
 static const char *TAG = "wifi_scan";
 
-// WiFi scan state
+#define WIFI_SCAN_MAX_RESULTS       20
+#define WIFI_SCAN_BSSID_STR_SIZE    18
+
 static struct {
     bool scan_in_progress;
     bool scan_completed;
@@ -23,32 +26,64 @@ static struct {
     .last_scan_result = ESP_OK
 };
 
-// WiFi scan mutex for thread safety
 static SemaphoreHandle_t wifi_scan_mutex = NULL;
 
-esp_err_t wifi_scan_init(void)
+static void wifi_scan_reset_state(void)
 {
-    if (wifi_scan_mutex == NULL) {
-        wifi_scan_mutex = xSemaphoreCreateMutex();
-        if (wifi_scan_mutex == NULL) {
-            ESP_LOGE(TAG, "Failed to create WiFi scan mutex");
-            return ESP_FAIL;
-        }
-    }
-    
-    // Reset scan state
     wifi_scan_state.scan_in_progress = false;
     wifi_scan_state.scan_completed = false;
     wifi_scan_state.ap_count = 0;
     wifi_scan_state.last_scan_result = ESP_OK;
-    
+}
+
+static void wifi_scan_prepare_for_start(void)
+{
+    wifi_scan_state.scan_in_progress = true;
+    wifi_scan_state.scan_completed = false;
+    wifi_scan_state.ap_count = 0;
+    // Keep last_scan_result as is - don't reset it here
+}
+
+static void wifi_scan_handle_start_failure(esp_err_t error)
+{
+    wifi_scan_state.scan_in_progress = false;
+    wifi_scan_state.last_scan_result = error;
+}
+
+static void wifi_scan_handle_completion(uint16_t ap_count)
+{
+    wifi_scan_state.scan_completed = true;
+    wifi_scan_state.scan_in_progress = false;
+    wifi_scan_state.ap_count = ap_count;
+    wifi_scan_state.last_scan_result = ESP_OK;
+}
+
+esp_err_t wifi_scan_init(void)
+{
+    if (wifi_scan_mutex != NULL) {
+        ESP_LOGW(TAG, "WiFi scan already initialized");
+        return ESP_OK;  // Don't fail, just warn
+    }
+
+    wifi_scan_mutex = xSemaphoreCreateMutex();
+    if (wifi_scan_mutex == NULL) {
+        ESP_LOGE(TAG, "Failed to create WiFi scan mutex");
+        return ESP_FAIL;
+    }
+
+    wifi_scan_reset_state();
+
     ESP_LOGI(TAG, "WiFi scan module initialized");
     return ESP_OK;
 }
 
-esp_err_t wifi_scan_start(void)
+static esp_err_t wifi_scan_start(void)
 {
-    // Take mutex for thread safety
+    if (wifi_scan_mutex == NULL) {
+        ESP_LOGE(TAG, "WiFi scan not initialized");
+        return ESP_FAIL;
+    }
+
     if (xSemaphoreTake(wifi_scan_mutex, pdMS_TO_TICKS(1000)) != pdTRUE) {
         ESP_LOGE(TAG, "Failed to take WiFi scan mutex");
         return ESP_FAIL;
@@ -56,19 +91,14 @@ esp_err_t wifi_scan_start(void)
 
     esp_err_t ret = ESP_OK;
 
-    // Check if scan is already in progress
     if (wifi_scan_state.scan_in_progress) {
         ESP_LOGW(TAG, "WiFi scan already in progress");
         ret = ESP_FAIL;
         goto exit;
     }
 
-    // Reset scan state
-    wifi_scan_state.scan_in_progress = true;
-    wifi_scan_state.scan_completed = false;
-    wifi_scan_state.ap_count = 0;
+    wifi_scan_prepare_for_start();
 
-    // Start WiFi scan (non-blocking)
     wifi_scan_config_t scan_config = {
         .ssid = NULL,
         .bssid = NULL,
@@ -83,11 +113,11 @@ esp_err_t wifi_scan_start(void)
         }
     };
 
-    esp_err_t scan_result = esp_wifi_scan_start(&scan_config, false); // false = non-blocking
+    esp_err_t scan_result = esp_wifi_scan_start(&scan_config, false);
+
     if (scan_result != ESP_OK) {
         ESP_LOGE(TAG, "Failed to start WiFi scan: 0x%x", scan_result);
-        wifi_scan_state.scan_in_progress = false;
-        wifi_scan_state.last_scan_result = scan_result;
+        wifi_scan_handle_start_failure(scan_result);
         ret = ESP_FAIL;
     } else {
         ESP_LOGI(TAG, "WiFi scan started successfully");
@@ -98,31 +128,36 @@ exit:
     return ret;
 }
 
-bool wifi_scan_is_in_progress(void)
+static bool wifi_scan_is_in_progress(void)
 {
     if (xSemaphoreTake(wifi_scan_mutex, pdMS_TO_TICKS(100)) != pdTRUE) {
         return false;
     }
-    
+
     bool in_progress = wifi_scan_state.scan_in_progress;
     xSemaphoreGive(wifi_scan_mutex);
     return in_progress;
 }
 
-bool wifi_scan_is_completed(void)
+static bool wifi_scan_is_completed(void)
 {
     if (xSemaphoreTake(wifi_scan_mutex, pdMS_TO_TICKS(100)) != pdTRUE) {
         return false;
     }
-    
+
     bool completed = wifi_scan_state.scan_completed;
     xSemaphoreGive(wifi_scan_mutex);
     return completed;
 }
 
-esp_err_t wifi_scan_get_results_json(cJSON **results_json)
+static esp_err_t wifi_scan_get_results_json(cJSON **results_json)
 {
     if (results_json == NULL) {
+        return ESP_FAIL;
+    }
+
+    if (wifi_scan_mutex == NULL) {
+        ESP_LOGE(TAG, "WiFi scan not initialized");
         return ESP_FAIL;
     }
 
@@ -149,11 +184,11 @@ esp_err_t wifi_scan_get_results_json(cJSON **results_json)
 
             cJSON_AddStringToObject(ap_json, "ssid", (const char *)wifi_scan_state.ap_records[i].ssid);
             cJSON_AddNumberToObject(ap_json, "rssi", wifi_scan_state.ap_records[i].rssi);
-            
+
             char bssid_str[WIFI_SCAN_BSSID_STR_SIZE];
             snprintf(bssid_str, sizeof(bssid_str), "%02X:%02X:%02X:%02X:%02X:%02X",
-                     wifi_scan_state.ap_records[i].bssid[0], wifi_scan_state.ap_records[i].bssid[1], 
-                     wifi_scan_state.ap_records[i].bssid[2], wifi_scan_state.ap_records[i].bssid[3], 
+                     wifi_scan_state.ap_records[i].bssid[0], wifi_scan_state.ap_records[i].bssid[1],
+                     wifi_scan_state.ap_records[i].bssid[2], wifi_scan_state.ap_records[i].bssid[3],
                      wifi_scan_state.ap_records[i].bssid[4], wifi_scan_state.ap_records[i].bssid[5]);
             cJSON_AddStringToObject(ap_json, "bssid", bssid_str);
             cJSON_AddNumberToObject(ap_json, "channel", wifi_scan_state.ap_records[i].primary);
@@ -166,9 +201,14 @@ esp_err_t wifi_scan_get_results_json(cJSON **results_json)
     return ESP_OK;
 }
 
-esp_err_t wifi_scan_get_status_json(cJSON **status_json)
+static esp_err_t wifi_scan_get_status_json(cJSON **status_json)
 {
     if (status_json == NULL) {
+        return ESP_FAIL;
+    }
+
+    if (wifi_scan_mutex == NULL) {
+        ESP_LOGE(TAG, "WiFi scan not initialized");
         return ESP_FAIL;
     }
 
@@ -189,16 +229,12 @@ esp_err_t wifi_scan_get_status_json(cJSON **status_json)
         // Try to get results to see if scan completed
         uint16_t ap_count = WIFI_SCAN_MAX_RESULTS;
         esp_err_t result = esp_wifi_scan_get_ap_records(&ap_count, wifi_scan_state.ap_records);
-        
+
         if (result == ESP_OK) {
-            wifi_scan_state.scan_completed = true;
-            wifi_scan_state.scan_in_progress = false;
-            wifi_scan_state.ap_count = ap_count;
-            wifi_scan_state.last_scan_result = ESP_OK;
+            wifi_scan_handle_completion(ap_count);
             ESP_LOGI(TAG, "WiFi scan completed, found %d networks", ap_count);
         } else if (result == ESP_ERR_WIFI_NOT_STARTED) {
-            wifi_scan_state.scan_in_progress = false;
-            wifi_scan_state.last_scan_result = result;
+            wifi_scan_handle_start_failure(result);
             ESP_LOGW(TAG, "WiFi scan failed: WiFi not started");
         }
     }
@@ -257,8 +293,8 @@ esp_err_t wifi_scan_results_handler(httpd_req_t *req)
 
     cJSON *resp_json = NULL;
     esp_err_t result = wifi_scan_get_status_json(&resp_json);
-    
-    if (result != ESP_OK || resp_json == NULL) {
+
+    if ((result != ESP_OK) || (resp_json == NULL)) {
         ESP_LOGE(TAG, "Failed to get scan status");
         return json_utils_send_error(req, "Failed to get scan status");
     }
@@ -266,7 +302,7 @@ esp_err_t wifi_scan_results_handler(httpd_req_t *req)
     // Add scan results if completed
     if (wifi_scan_is_completed()) {
         cJSON *networks = NULL;
-        if (wifi_scan_get_results_json(&networks) == ESP_OK && networks != NULL) {
+        if ((wifi_scan_get_results_json(&networks) == ESP_OK) && (networks != NULL)) {
             cJSON_AddItemToObject(resp_json, "networks", networks);
         }
     }
